backend/fastapi_ml/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.6.0
python-dotenv==1.0.1
requests==2.32.3

# ML dependencies for food recommendation system
tensorflow>=2.20.0  # Latest version compatible with Python 3.13
pyarrow>=18.0.0  # For parquet file support
scipy>=1.14.0
matplotlib>=3.9.0
seaborn>=0.13.0

# Development dependencies
pytest==7.4.3
httpx==0.25.2
